+++ POD/src/GPSProcessing/PPPSolverLEO.cpp
@@ -362,7 +362,7 @@ covariance matrix.");
 
             Vector<double> prefitC(gData.getVectorOfTypeID(defaultEqDef.header));
             Vector<double> prefitL(gData.getVectorOfTypeID(TypeID::prefitL));
-            for (int i = 0; i < numCurrentSV; i++)
+            for (size_t i = 0; i < numCurrentSV; i++)
             {
                 measVector(i) = prefitC(i);
                 measVector(numCurrentSV + i) = prefitL(i);
@@ -377,14 +377,14 @@ covariance matrix.");
             satTypeValueMap dummy(gData.body.extractTypeID(TypeID::weight));
 
             // Check if weights match
-            if (dummy.numSats() == (size_t)numCurrentSV)
+            if (dummy.numSats() == numCurrentSV)
             {
 
                 // If we have weights information, let's load it
                 Vector<double>
                     weightsVector(gData.getVectorOfTypeID(TypeID::weight));
 
-                for (int i = 0; i < numCurrentSV; i++)
+                for (size_t i = 0; i < numCurrentSV; i++)
                 {
 
                     rMatrix(i, i) = weightsVector(i);
@@ -398,7 +398,7 @@ covariance matrix.");
             {
 
                 // If weights don't match, assign generic weights
-                for (int i = 0; i < numCurrentSV; i++)
+                for (size_t i = 0; i < numCurrentSV; i++)
                 {
                     rMatrix(i, i) = 1.0;
 
@@ -419,11 +419,11 @@ covariance matrix.");
             Matrix<double> dMatrix(gData.body.getMatrixOfTypes(defaultEqDef.body));
 
             // Let's fill 'hMatrix'
-            for (int i = 0; i < numCurrentSV; i++)
+            for (size_t i = 0; i < numCurrentSV; i++)
             {
 
                 // First, fill the coefficients related to tropo, coord and clock
-                for (int j = 0; j < numVar; j++)
+                for (size_t j = 0; j < numVar; j++)
                 {
 
                     hMatrix(i, j) = dMatrix(i, j);
@@ -670,7 +670,7 @@ covariance matrix.");
                // Now we have to add the new values to the data structure
             Vector<double> postfitCode(numCurrentSV, 0.0);
             Vector<double> postfitPhase(numCurrentSV, 0.0);
-            for (int i = 0; i < numCurrentSV; i++)
+            for (size_t i = 0; i < numCurrentSV; i++)
             {
                 postfitCode(i) = postfitResiduals(i);
                 postfitPhase(i) = postfitResiduals(i + numCurrentSV);
